optimise: split v10 to measure only 1 optimisation per version
