+++ optimise/timings.md
@@ -165,6 +165,17 @@ Timing summary
                      'eta_MAX_iter': 4531.594329423113,
                      'ts_ratio': 10.265947010609201}},
  'v09': {'pypy3': {1048576: {100: 1.43,
+                             1000: 13.37,
+                             10000: 133.27,
+                             100000: 1342.05},
+                   'eta_MAXN': 'N/A',
+                   'eta_MAX_iter': 1342.05,
+                   'ts_ratio': 9.795882373044352},
+         'python3': {1048576: {100: 3.26, 1000: 33.89, 10000: 343.85},
+                     'eta_MAXN': 'N/A',
+                     'eta_MAX_iter': 3531.6431722217494,
+                     'ts_ratio': 10.270883153182346}},
+ 'v10': {'pypy3': {1048576: {100: 1.43,
                              1000: 13.34,
                              10000: 133.25,
                              100000: 1340.49},
@@ -175,7 +186,7 @@ Timing summary
                      'eta_MAXN': 'N/A',
                      'eta_MAX_iter': 3513.287494355102,
                      'ts_ratio': 10.252983990997205}},
- 'v10': {'pypy3': {1048576: {100: 3.84, 1000: 35.01, 10000: 327.67},
+ 'v11': {'pypy3': {1048576: {100: 3.84, 1000: 35.01, 10000: 327.67},
                    'eta_MAXN': 'N/A',
                    'eta_MAX_iter': 3027.0995740167987,
                    'ts_ratio': 9.238256703441873},
@@ -283,13 +294,19 @@ Speedup summary
          'total_python3': 42411655589441.32,
          'total_vs_python3_pypy3': 142409059092872.12,
          'vs_python3_pypy3': 3.3577811833482367},
- 'v09': {'prev_pypy3': 1.0067811024326925,
-         'prev_python3': 1.2898444367858177,
+ 'v09': {'prev_pypy3': 1.0056108192690287,
+         'prev_python3': 1.283140484029222,
+         'total_pypy3': 18609363346503.344,
+         'total_python3': 54420112281516.4,
+         'total_vs_python3_pypy3': 143208090585714.66,
+         'vs_python3_pypy3': 2.631528759898476},
+ 'v10': {'prev_pypy3': 1.0011637535528053,
+         'prev_python3': 1.0052246444095851,
          'total_pypy3': 18631020059213.28,
          'total_python3': 54704438016917.016,
          'total_vs_python3_pypy3': 143374749509924.25,
          'vs_python3_pypy3': 2.6208979510142574},
- 'v10': {'prev_pypy3': 0.4428298333844505,
+ 'v11': {'prev_pypy3': 0.4428298333844505,
          'prev_python3': 0.8175693611799487,
          'total_pypy3': 8250371508603.772,
          'total_python3': 44724672443198.95,
